forex_app_backend/microservicio_divisas/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
pytest==7.4.3